Update READMEs and conversion script
